requirements.txt
amqp==2.2.1
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.9.1
Automat==0.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
channels==1.1.6
chardet==3.0.4
constantly==15.1.0
daphne==1.3.0
Django==1.11.5
django-channels==0.7.0
gunicorn==19.7.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
kombu==4.1.0
msgpack-python==0.4.8
mysqlclient==1.3.12
oauthlib==2.0.7
parse==1.8.2
psycopg2==2.7.3.1
PyJWT==1.6.1
PySocks==1.6.8
python-decouple==3.1
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
twilio==6.11.0
Twisted==17.5.0
txaio==2.8.2
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
zope.interface==4.4.2
